+++ generator/lib/builder.dart
@@ -2,4 +2,4 @@ import 'package:build/build.dart';
 import 'src/generator.dart';
 
 Builder retrofitBuilder(BuilderOptions options) =>
-    generatorFactoryBuilder(header: options.config["header"] as String);
+    generatorFactoryBuilder(options);