+++ generator/lib/src/generator.dart
@@ -14,6 +14,17 @@ import 'package:tuple/tuple.dart';
 
 import 'package:retrofit/retrofit.dart' as retrofit;
 
+class RetrofitOptions {
+  final bool autoCastResponse;
+
+  RetrofitOptions({this.autoCastResponse});
+
+  RetrofitOptions.fromOptions([BuilderOptions options])
+      : autoCastResponse =
+            (options?.config['auto_cast_response']?.toString() ?? 'true') ==
+                'true';
+}
+
 class RetrofitGenerator extends GeneratorForAnnotation<retrofit.RestApi> {
   static const String _baseUrlVar = 'baseUrl';
   static const _queryParamsVar = "queryParameters";
@@ -31,6 +42,14 @@ class RetrofitGenerator extends GeneratorForAnnotation<retrofit.RestApi> {
 
   var _customBaseUrl = false;
 
+  /// Global options sepcefied in the `build.yaml`
+  final RetrofitOptions globalOptions;
+
+  RetrofitGenerator(this.globalOptions);
+
+  /// Annotation details for [RestApi]
+  retrofit.RestApi clientAnnotation;
+
   @override
   String generateForAnnotatedElement(
       Element element, ConstantReader annotation, BuildStep buildStep) {
@@ -47,7 +66,11 @@ class RetrofitGenerator extends GeneratorForAnnotation<retrofit.RestApi> {
   String _implementClass(ClassElement element, ConstantReader annotation) {
     final className = element.name;
     final name = '_$className';
-    final baseUrl = annotation?.peek(_baseUrlVar)?.stringValue ?? '';
+    clientAnnotation = retrofit.RestApi(
+      autoCastResponse: (annotation?.peek('autoCastResponse')?.boolValue),
+      baseUrl: (annotation?.peek(_baseUrlVar)?.stringValue ?? ''),
+    );
+    final baseUrl = clientAnnotation.baseUrl;
     _customBaseUrl = _isValidBaseUrl(baseUrl);
     final classBuilder = new Class((c) {
       c
@@ -236,6 +259,9 @@ class RetrofitGenerator extends GeneratorForAnnotation<retrofit.RestApi> {
     for (var parameter in m.parameters.where((p) =>
         p.isRequiredNamed ||
         p.isRequiredPositional ||
+        // TODO: remove this after requried syntax is available https://github.com/dart-lang/language/issues/15
+        // ignore: deprecated_member_use
+        p.isRequired ||
         p.metadata.firstWhere((meta) => meta.isRequired, orElse: () => null) !=
             null)) {
       blocks.add(Code(
@@ -270,15 +296,35 @@ class RetrofitGenerator extends GeneratorForAnnotation<retrofit.RestApi> {
     namedArguments[_dataVar] = refer(_localDataVar);
 
     final cancelToken = _getAnnotation(m, retrofit.CancelRequest);
-    if ( cancelToken != null ) namedArguments[_cancelToken] = refer(cancelToken.item1.displayName);
+    if (cancelToken != null)
+      namedArguments[_cancelToken] = refer(cancelToken.item1.displayName);
 
     final sendProgress = _getAnnotation(m, retrofit.SendProgress);
-    if ( sendProgress != null ) namedArguments[_onSendProgress] = refer(sendProgress.item1.displayName);
+    if (sendProgress != null)
+      namedArguments[_onSendProgress] = refer(sendProgress.item1.displayName);
 
     final receiveProgress = _getAnnotation(m, retrofit.ReceiveProgress);
-    if ( receiveProgress != null ) namedArguments[_onReceiveProgress] = refer(receiveProgress.item1.displayName);
+    if (receiveProgress != null)
+      namedArguments[_onReceiveProgress] =
+          refer(receiveProgress.item1.displayName);
 
     final returnType = _getResponseType(m.returnType);
+
+    final autoCastResponse = (globalOptions.autoCastResponse ??
+        (clientAnnotation.autoCastResponse ?? true) ??
+        (httpMehod.peek('autoCastResponse')?.boolValue ?? true));
+
+    /// If autoCastResponse is false, return the response as it is
+    if (!autoCastResponse) {
+      blocks.add(
+        refer("$_dioVar.request")
+            .call([path], namedArguments)
+            .returned
+            .statement,
+      );
+      return Block.of(blocks);
+    }
+
     if (returnType == null || "void" == returnType.toString()) {
       blocks.add(
         refer("await $_dioVar.request")
@@ -299,7 +345,8 @@ class RetrofitGenerator extends GeneratorForAnnotation<retrofit.RestApi> {
                 .assignFinal(_resultVar, refer("Response<List<dynamic>>"))
                 .statement,
           );
-          blocks.add(Code("final value = $_resultVar.data.cast<$innerReturnType>();"));
+          blocks.add(
+              Code("final value = $_resultVar.data.cast<$innerReturnType>();"));
         } else {
           blocks.add(
             refer("await $_dioVar.request")
@@ -415,17 +462,22 @@ class RetrofitGenerator extends GeneratorForAnnotation<retrofit.RestApi> {
         ]).statement);
       } else if (_bodyName.type.element is ClassElement) {
         final ele = _bodyName.type.element as ClassElement;
-        final toJson = ele.methods.firstWhere((i) => i.displayName == "toJson");
+        final toJson = ele.methods
+            .firstWhere((i) => i.displayName == "toJson", orElse: () => null);
         if (toJson == null) {
-          log.severe(
-              "${_bodyName.type} must provide a `toJson()` method which return a Map.");
+          log.warning(
+              "${_bodyName.type} must provide a `toJson()` method which return a Map.\n"
+              "It is programmer's responsibility to make sure the ${_bodyName.type} is properly serialized");
+          blocks.add(
+              refer(_bodyName.displayName).assignFinal(_dataVar).statement);
+        } else {
+          blocks.add(literalMap({}, refer("String"), refer("dynamic"))
+              .assignFinal(_dataVar)
+              .statement);
+          blocks.add(refer("$_dataVar.addAll").call([
+            refer("${_bodyName.displayName}.toJson() ?? <String,dynamic>{}")
+          ]).statement);
         }
-        blocks.add(literalMap({}, refer("String"), refer("dynamic"))
-            .assignFinal(_dataVar)
-            .statement);
-        blocks.add(refer("$_dataVar.addAll").call([
-          refer("${_bodyName.displayName}.toJson() ?? <String,dynamic>{}")
-        ]).statement);
       } else {
         /// @Body annotations with no type are assinged as is
         blocks
@@ -522,8 +574,10 @@ class RetrofitGenerator extends GeneratorForAnnotation<retrofit.RestApi> {
   }
 }
 
-Builder generatorFactoryBuilder({String header}) =>
-    new SharedPartBuilder([new RetrofitGenerator()], "retrofit");
+Builder generatorFactoryBuilder(BuilderOptions options) =>
+    new SharedPartBuilder(
+        [new RetrofitGenerator(RetrofitOptions.fromOptions(options))],
+        "retrofit");
 
 /// Returns `$revived($args $kwargs)`, this won't have ending semi-colon (`;`).
 /// [object] must not be null.